Reorganise more files into packages
