+++ src/main/java/matcha/Matcha.java
@@ -1,6 +1,10 @@
 package matcha;
 import matcha.command.Command;
 import matcha.exception.MatchaException;
+import matcha.parser.Parser;
+import matcha.storage.Storage;
+import matcha.tasklist.TaskList;
+import matcha.ui.Ui;
 
 public class Matcha {
     private static final String FILE_PATH = "./data/matcha.txt";
