+++ src/main/java/matcha/command/Command.java
@@ -1,8 +1,8 @@
 package matcha.command;
 import matcha.exception.MatchaException;
-import matcha.Storage;
-import matcha.TaskList;
-import matcha.Ui;
+import matcha.storage.Storage;
+import matcha.tasklist.TaskList;
+import matcha.ui.Ui;
 
 public abstract class Command {
