+++ src/main/java/matcha/command/ExitCommand.java
@@ -1,7 +1,7 @@
 package matcha.command;
-import matcha.Storage;
-import matcha.TaskList;
-import matcha.Ui;
+import matcha.storage.Storage;
+import matcha.tasklist.TaskList;
+import matcha.ui.Ui;
 
 public class ExitCommand extends Command{
     @Override
