+++ src/main/java/matcha/command/ListTaskCommand.java
@@ -1,7 +1,7 @@
 package matcha.command;
-import matcha.TaskList;
-import matcha.Storage;
-import matcha.Ui;
+import matcha.tasklist.TaskList;
+import matcha.storage.Storage;
+import matcha.ui.Ui;
 
 public class ListTaskCommand extends Command{
     @Override
