+++ src/main/java/matcha/storage/Storage.java
@@ -1,4 +1,4 @@
-package matcha;
+package matcha.storage;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
